refactor(shared): extract realmToSlug + parseInGameName + DEFAULT_REGION to shared
Move three shared helpers into a single source of truth in
`@mythicplus/shared` so the activity frontend and the Firebase functions
agree on slug normalization, name parsing, and the default region.

Consumer sites updated:
- activity/src/lib/currentCharacter.ts now re-exports from shared (preserves
  backward-compat for activity callers like useIdentity that import via
  this module).
- activity/src/components/RoleEditor.tsx imports DEFAULT_REGION (and
  continues to use parseInGameName, which is now sourced from shared).
- activity/src/hooks/useDungeonSuggestions.ts drops its local
  realmToSlug/parseInGameName/DEFAULT_REGION duplicates.
- packages/functions/src/refreshCharacterMedia.ts drops its local copies
  and switches the parse-result field from `realm` to `realmSlug` to
  match the canonical shape.

Verified `./scripts/verify-ts.sh` and `cd activity && npm run typecheck`
both pass; activity unit tests (incl. currentCharacter.test.ts) all green.
